Release 6

+++ 3_Implementation/src/Derivatives.c
@@ -1,4 +1,5 @@
-int derivative(float c, float x ,float n)
+#include<math.h>
+float derivative(float c, float x ,float n)
 {
 	float powerval = pow(x, n-1);
 	return n * c * powerval;
